% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cat_results}
\alias{cat_results}
\title{Categorical Simulation Results}
\format{
4000 rows and 9 variables. Relevant variables include:
\describe{
\item{\code{Estimate.P}}{ Estimate of Atrazine presence}
\item{\code{StdError.P}}{ Standard error of Atrazine presence}
\item{\code{LCB95Pct.P}}{ Lower confidence bound of Atrazine presence}
\item{\code{UCB95Pct.P}}{ Upper confidence bound of Atrazine presence}
\item{\code{spb}}{ Spatial balance as measured by Pielou's evenness index}
\item{\code{algorithm}}{ Sampling algorithm (GRTS or SRS)}
\item{\code{true_mean}}{ True mean Atrazine presence}
\item{\code{cover}}{ Was the true mean contained in the confidence interval (1 = yes, 0 = no)}
\item{\code{seed}}{ The simulation seed (which equals the trial number)}
}
}
\usage{
cat_results
}
\description{
An data frame of results for the Atrazine simulation.
}
\keyword{datasets}
