% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cont_results}
\alias{cont_results}
\title{Continuous Simulation Results}
\format{
4000 rows and 9 variables. Relevant variables include:
\describe{
\item{\code{Estimate}}{ Estimate of BMMI}
\item{\code{StdError}}{ Standard error of BMMI}
\item{\code{LCB95Pct}}{ Lower confidence bound of BMMI}
\item{\code{UCB95Pct}}{ Upper confidence bound of BMMI}
\item{\code{spb}}{ Spatial balance as measured by Pielou's evenness index}
\item{\code{algorithm}}{ Sampling algorithm (GRTS or SRS)}
\item{\code{true_mean}}{ True mean BMMI}
\item{\code{cover}}{ Was the true mean contained in the confidence interval (1 = yes, 0 = no)}
\item{\code{seed}}{ The simulation seed (which equals the trial number)}
}
}
\usage{
cont_results
}
\description{
An data frame of results for the BMMI simulation.
}
\keyword{datasets}
