% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cont_summaries}
\alias{cont_summaries}
\title{Continuous Simulation Summary}
\format{
2 rows and 6 variables. Relevant variables include:
\describe{
\item{\code{algorithm}}{ Sampling algorithm (GRTS or SRS)}
\item{\code{mbias}}{ Mean bias}
\item{\code{mstderr}}{ Mean standard error}
\item{\code{rmse}}{ Root-Mean-Squared Error}
\item{\code{coverage}}{ Coverage rate}
\item{\code{mspb}}{ Mean spatial balance as measured by Pielou's evenness index}
}
}
\usage{
cont_summaries
}
\description{
An data frame of summaries for the BMMI simulation.
}
\keyword{datasets}
