% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NLA12}
\alias{NLA12}
\title{NLA 2012 Data}
\format{
1030 rows and 3 variables:
\describe{
\item{\code{AP}}{ Atrazine presence. Takes on values Yes (\code{"Y"}) and
No (\code{"No"}).}
\item{\code{BMMI}}{ Benthic macroinvertebrate multi-metric index value.}
\item{\code{geometry}}{ POINT geometry using the NAD83 / Conus Albers
coordinate reference system (EPSG: 5070).}
}
}
\usage{
NLA12
}
\description{
An \code{sf} POINT object of data collected at 1030 lakes during EPA's 2012
National Lakes Assessment.
}
\keyword{datasets}
