% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NLA12_keyvariables}
\alias{NLA12_keyvariables}
\title{NLA 2012 Raw Data}
\format{
1138 rows and 49 variables. Relevant variables include:
\describe{
\item{\code{AP}}{ Atrazine presence. Takes on values Yes (\code{"Y"}) and
No (\code{"No"}).}
\item{\code{BMMI}}{ Benthic macroinvertebrate multi-metric index value.}
\item{\code{INDEX_LON_DD}}{ Latitude coordinate using the NAD83
coordinate reference system (EPSG: 4269).}
\item{\code{INDEX_LAT_DD}}{ Longitude coordinate using the NAD83
coordinate reference system (EPSG: 4269).}
}
}
\usage{
NLA12_keyvariables
}
\description{
An \code{sf} POINT object of data collected at 1138 lakes during EPA's 2012
National Lakes Assessment. The data are available at
https://www.epa.gov/sites/production/files/2021-04/nla12_keyvariables.csv
}
\keyword{datasets}
